src/agent/requirements.txt
# Strands agent dependencies with OpenAI support
strands-agents[openai]==1.32.0
strands-agents-tools==0.5.2
bedrock-agentcore==1.4.7
mcp==1.26.0
PyJWT[crypto]==2.12.1
